requirements.txt
praw>=7.7.0
openai>=1.3.0
PyYAML>=6.0
argparse>=1.4.0
requests>=2.31.0
google-generativeai>=0.3.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.8
anthropic>=0.5.0
